+++ src/components/Column/index.ts
@@ -1,0 +1,3 @@
+import RvColumn from "./RvColumn";
+
+export { RvColumn };