+++ src/components/Column/style/index.ts
@@ -1,0 +1,3 @@
+import StyledColumn from "./StyledColumn";
+
+export { StyledColumn };