+++ src/components/Main/index.ts
@@ -1,0 +1,3 @@
+import RvMain from "./RvMain";
+
+export { RvMain };