+++ src/components/Row/index.ts
@@ -1,0 +1,3 @@
+import RvRow from "./RvRow";
+
+export { RvRow };