+++ src/components/Stat/index.ts
@@ -1,0 +1,4 @@
+import RvStat from "./RvStat";
+import RvStatHeader from "./RvStatHeader";
+
+export { RvStat, RvStatHeader };