+++ src/types/components/Props.ts
@@ -1,0 +1,6 @@
+import type { ReactNode } from "react";
+
+export type BasePropsType<T> = {
+	[props: string]: object | ReactNode | T;
+	children?: ReactNode;
+};