+++ src/types/components/Stat.ts
@@ -1,0 +1,5 @@
+import type { RvCardFCType, RvCardHeaderFCType } from "./Card";
+
+export type RvStatFCType = RvCardFCType;
+
+export type RvStatHeaderFCType = RvCardHeaderFCType;